requirements.txt
Flask
flask-cors
gspread
google-auth
gunicorn
openai
werkzeug.utils